Sort reference points by name when inserting a new refpoint in a panorama
With lots of reference points, the current sorting by distance is really
not usable.  Sorting by name allows to more easily find a given refpoint
in the list.

While we're at it, display the distance in the dropdown menu so that it's
easier to check if it's the right refpoint.