Add a Django command to separate ground altitude from height
It uses the API to estimate the ground altitude for each point, and then
sets the remaining altitude to "height above ground".  The total altitude
(ground altitude + height above ground) remains unchanged.

To see what would happen with your data without saving the result, run:

    ./manage.py fix_ground_altitude --dry-run

If you're happy with the result, run it for real:

    ./manage.py fix_ground_altitude